Cleanup source codes for remaining applications.
