+++ CaffeGoogLeNet/googleNet.cpp
@@ -32,157 +32,53 @@
 #include <string>
 #include <vector>
 
-#include "imagenet_1000_categories.h"
-
 #include "CaffeGoogLeNet_gen.h"
 #include "util_draw.h"
 #include "util_input.h"
-CCaffeGoogLeNet network;
-
-#define FILENAME_WEIGHTS "CaffeGoogLeNet_weights.bin"
+#include "demo_common.h"
+#include "imagenet_1000_categories.h"
 
 using namespace std;
 using namespace dmp;
 using namespace util;
 
-#define SCREEN_W (dmp::util::get_screen_width())
-#define SCREEN_H (dmp::util::get_screen_height())
+#define SCREEN_W (get_screen_width())
+#define SCREEN_H (get_screen_height())
 
 #define IMAGE_W 224
 #define IMAGE_H 224
 
+#define FILENAME_WEIGHTS "CaffeGoogLeNet_weights.bin"
+
+// Define CNN network model object
+CCaffeGoogLeNet network;
+
+// Categories strings
 std::vector<std::string> catstr_vec(categories, categories + 1000);
 
-uint32_t fc = 0;
-
+// Buffer for decoded image data
 uint32_t imgView[IMAGE_W * IMAGE_H];
+// Buffer for pre-processed image data
 __fp16 imgProc[IMAGE_W * IMAGE_H * 3];
 
-// 2ND THREAD FOR HW CONTROL
-
-volatile uint64_t sync_cnn_in = 0;
-volatile uint64_t sync_cnn_out = 0;
-
-volatile int conv_time_tot = 0;
-volatile int fc_time_tot = 0;
-
-volatile bool g_should_stop = false;
-
-void* hwacc_thread_func(void* targ) {
-  while (!g_should_stop) {
-    if (sync_cnn_in == sync_cnn_out) {
-      usleep(1000);  // sleep 1 ms
-      continue;
-    }
-
-    network.RunNetwork();
-
-    conv_time_tot = network.get_conv_usec();
-    fc_time_tot = network.get_fc_usec();
-
-    sync_cnn_out++;
-  }
-  return NULL;
-}
-
-void print_demo_title(COverlayRGB &bg_overlay)
-{
-  unsigned text_size = 30;
-  string font_file = "font/NotoSerif-Black.ttf";
-  string text = "CNN - GoogleNet";
-  unsigned w = 0;
-  unsigned h = 0;
-
-  COverlayRGB::calculate_boundary_text_with_font(font_file, text, text_size, w, h);
-  int x = ((SCREEN_W - w) / 2);
-  int y = 25;
-  COverlayRGB bg_text(SCREEN_W, SCREEN_H);
-  bg_text.alloc_mem_overlay(w, h);
-  bg_text.copy_overlay(bg_overlay, x, y);
-  bg_text.set_text_with_font(font_file, text, 0, 3*h/4, text_size, 0x00ffffff);
-  bg_text.print_to_display(x, y);
-
-
-  text = "FPGA Demonstration";
-  COverlayRGB::calculate_boundary_text_with_font(font_file, text, text_size, w, h);
-  x = ((SCREEN_W - w) / 2);
-  y = 65;
-  bg_text.delete_overlay();
-  bg_text.alloc_mem_overlay(w, h);
-  bg_text.copy_overlay(bg_overlay, x, y);
-  bg_text.set_text_with_font(font_file, text, 0, 3*h/4, text_size, 0x00ffffff);
-  bg_text.print_to_display(x, y);
-
-  text_size = 11;
-  text = "Copyright 2018. Digital Media Professionals Inc.";
-  COverlayRGB::calculate_boundary_text_with_font(font_file, text, text_size, w, h);
-  x = 5;
-  y = SCREEN_H - 20;
-  bg_text.delete_overlay();
-  bg_text.alloc_mem_overlay(w, h);
-  bg_text.copy_overlay(bg_overlay, x, y);
-  bg_text.set_text_with_font(font_file, text, 0, 3*h/4, text_size, 0x00ffffff);
-  bg_text.print_to_display(x, y);
-}
-
 int main(int argc, char** argv) {
-  if (!dmp::util::init_fb()) {
-    fprintf(stderr, "dmp::util::init_fb() failed\n");
+  // Initialize FB
+  if (!init_fb()) {
+    cout << "init_fb() failed." << endl;
     return 1;
   }
 
-  const std::string input_image_path = "./images/";
-  const std::vector<std::string> input_image_suffix = {".jpg", ".jpeg", ".JPG",
-                                                       ".JPEG"};
-
-  vector<string> image_names =
-      dmp::util::get_input_image_names(input_image_path, input_image_suffix);
+  // Get input images filenames
+  vector<string> image_names;
+  get_jpeg_image_names("./images/", image_names);
   int num_images = image_names.size();
   if (num_images == 0) {
     cout << "No input images." << endl;
     return 1;
   }
 
-  std::vector<float> networkOutput;
-
-  int my_number = 0;
-  bool has_democonf = false;
-  if (argc >= 2) {
-    my_number = atoi(argv[1]);
-    has_democonf = true;
-  }
-  vector<pair<int, string> > democonf;
-  int democonf_sel = 0;  // currencly selected element index (0 .. num-1)
-  int democonf_num = 0;  // number of elements in democonf
-  int democonf_string_max = -1;
-  bool democonf_display = false;
-  if (has_democonf) {
-    ifstream democonf_file("democonf.txt");
-    int count = 0;
-    int c_int;
-    string c_string;
-    while (democonf_file >> c_int >> c_string) {
-      democonf.push_back(make_pair(c_int, c_string));
-      int c_string_size = c_string.size();
-      if (c_string_size > democonf_string_max)
-        democonf_string_max = c_string_size;
-      if (c_int == my_number) democonf_sel = count;
-      count++;
-    }
-    democonf_num = count;
-  }
-
-  COverlayRGB bg_overlay(SCREEN_W, SCREEN_H);
-  bg_overlay.alloc_mem_overlay(SCREEN_W, SCREEN_H);
-  bg_overlay.load_ppm_img("fpgatitle");
-  COverlayRGB overlay_input(SCREEN_W, SCREEN_H);
-  overlay_input.alloc_mem_overlay(IMAGE_W, IMAGE_H);
-
+  // Initialize network object
   network.Verbose(0);
-  const bool dump_outputs = false;
-  if (dump_outputs) {
-    network.WantLayerOutputs();
-  }
   if (!network.Initialize()) {
     return -1;
   }
@@ -193,215 +89,71 @@ int main(int argc, char** argv) {
     return -1;
   }
 
+  // Get HW module frequency
   string conv_freq, fc_freq;
   conv_freq = std::to_string(network.get_dv_info().conv_freq);
   fc_freq = std::to_string(network.get_dv_info().fc_freq);
 
-  void* ddr_buf_a_cpu = network.get_network_input_addr_cpu();
+  // Create background and image overlay
+  COverlayRGB bg_overlay(SCREEN_W, SCREEN_H);
+  bg_overlay.alloc_mem_overlay(SCREEN_W, SCREEN_H);
+  bg_overlay.load_ppm_img("fpgatitle");
+  COverlayRGB overlay_input(SCREEN_W, SCREEN_H);
+  overlay_input.alloc_mem_overlay(IMAGE_W, IMAGE_H);
 
-  int exit_code = -1;
-
-  int image_nr = 0;
-
-  bool pause = false;
-
-  pthread_t hwacc_thread;
-  pthread_create(&hwacc_thread, NULL, hwacc_thread_func, NULL);
-
-  while (exit_code == -1) {
-    // Static Images
-    if (fc < 2) {
-      bg_overlay.print_to_display(0, 0);
-      print_demo_title(bg_overlay);
-      dmp::util::swap_buffer();
-      fc++;  // Frame Counter
-      continue;
-    }
-
-    if (sync_cnn_out == sync_cnn_in) {
-      if (sync_cnn_out != 0) {
-        // HW processing times
-        if (conv_time_tot != 0 && fc_time_tot != 0) {
-          string text = COverlayRGB::convert_time_to_text(
-                  "Convolution (" + conv_freq + " MHz HW ACC)      : ", conv_time_tot);
-          unsigned text_size = 14;
-
-          unsigned w = 0;
-          unsigned h = 0;
-          COverlayRGB::calculate_boundary_text(text, text_size, w, h);
-
-          int x = ((SCREEN_W - w) / 2);
-          int y = 7*SCREEN_H/8;
-
-          COverlayRGB overlay_time(SCREEN_W, SCREEN_H);
-          overlay_time.alloc_mem_overlay(w, h);
-          overlay_time.copy_overlay(bg_overlay,x, y);
-          overlay_time.set_text(0, 0, text, text_size, 0x00f4419d);
-          overlay_time.print_to_display(x, y);
-
-          text = COverlayRGB::convert_time_to_text(
-                    "Fully Connected (" + fc_freq + " MHz HW ACC) : ", fc_time_tot);
-          COverlayRGB::calculate_boundary_text(text, text_size, w, h);
-
-          y = 7*SCREEN_H/8 + 28;
-
-          COverlayRGB overlay_fc(SCREEN_W, SCREEN_H);
-          overlay_fc.alloc_mem_overlay(w, h);
-          overlay_fc.copy_overlay(bg_overlay,x, y);
-          overlay_fc.set_text(0, 0, text, text_size, 0x00f4419d);
-          overlay_fc.print_to_display(x, y);
-
-
-          text = COverlayRGB::convert_time_to_text(
-                    "Total Processing Time           : ", conv_time_tot + fc_time_tot);
-          COverlayRGB::calculate_boundary_text(text, text_size, w, h);
-
-          y = 7*SCREEN_H/8 + 56;
-
-          COverlayRGB overlay_processingtime(SCREEN_W, SCREEN_H);
-          overlay_processingtime.alloc_mem_overlay(w, h);
-          overlay_processingtime.copy_overlay(bg_overlay,x, y);
-          overlay_processingtime.set_text(0, 0, text, text_size, 0x00f4419d);
-          overlay_processingtime.print_to_display(x, y);
-        }
-
-        network.get_final_output(networkOutput);
-
-        int x = (SCREEN_W - IMAGE_W) / 2;
-        int y = (293 - 128) + 20;
-        overlay_input.print_to_display(x, y);
-
-        x = (SCREEN_W / 5);
-        y = (293 - 128) + IMAGE_W + 80;
-
-        print_result(catstr_vec, x, y,
-                    dmp::util::catrank(&networkOutput.front()), bg_overlay);
-
-        if (dump_outputs) {
-          const int n_layers = network.get_total_layer_count();
-          for (int i_layer = 0; i_layer < n_layers; ++i_layer) {
-            fpga_layer& layer = network.get_layer(i_layer);
-            char fnme[256];
-            snprintf(fnme, sizeof(fnme), "%02d.%dx%dx%d.bin", i_layer,
-                     layer.output_dim[0], layer.output_dim[1], layer.output_dim[2]);
-            FILE *fout = fopen(fnme, "wb");
-            if (!fout) {
-              fprintf(stderr, "fopen() failed for %s\n", fnme);
-              fflush(stderr);
-              return -1;
-            }
-            fwrite(layer.output.data(), sizeof(layer.output[0]), layer.output.size(), fout);
-            fclose(fout);
-            fprintf(stdout, "Saved %s\n", fnme);
-            fflush(stdout);
-          }
-          fflush(stderr);
-          break;
-        }
-
-        dmp::util::swap_buffer();
-        fc++;
-
-        int key = getchar();
-        switch (key) {
-          case 27:  // ESC
-          {
-            int next_key = getchar();
-            switch (next_key) {
-              case 91:  // there are more value to read: UP/DOWN/LEFT/RIGHT pressed
-                break;
-              case 79:  // F3 pressed
-                break;
-              default:  // nothing special was pressed, will exit
-                exit_code = 0;
-                break;
-            }
-            break;
-          }
-          case '3':  // exit demo with exit code of selected next demo
-            if (has_democonf) {
-              int sel_num = democonf[democonf_sel].first;
-              if (sel_num != my_number) {
-                exit_code = sel_num;
-              }
-              else {
-                exit_code = my_number;
-              }
-            }
-            break;
-
-          case '2':  // cycle through demo configuratom list
-            if (has_democonf) {
-              democonf_display = true;
-              if (democonf_sel == democonf_num - 1) {
-                democonf_sel = 0;
-              }
-              else {
-                democonf_sel++;
-              }
-            }
-            break;
-
-          case '1':
-          case 32:  // SPACE
-            pause = !pause;
-            break;
-        }
-      }
-
-      if (!pause) {
-        dmp::util::decode_jpg_file(input_image_path + image_names[image_nr],
-                                   imgView, IMAGE_W, IMAGE_H);
-        overlay_input.convert_to_overlay_pixel_format(imgView, IMAGE_W*IMAGE_H);
-        dmp::util::preproc_image(imgView, imgProc, IMAGE_W, IMAGE_H, -128.0,
-                                 -128.0, -128.0, 1.0, true);
-
-        if (image_nr == num_images - 1) {
-          image_nr = 0;
-        } else {
-          image_nr++;
-        }
-      }
-
-      // Copy image to FPGA memory
-      memcpy(ddr_buf_a_cpu, imgProc, IMAGE_W * IMAGE_H * 3 * 2);
-
-      if (dump_outputs) {
-        FILE *fout = fopen("input.bin", "wb");
-        if (fout) {
-          fwrite(ddr_buf_a_cpu, 2, IMAGE_W * IMAGE_H * 3, fout);
-          fclose(fout);
-          fprintf(stdout, "Saved input.bin\n");
-          fflush(stdout);
-        }
-      }
-
-      if (exit_code == -1)  // do not start new HW ACC runs if about to exit...
-        sync_cnn_in++;
-    }
-
-    if (democonf_display) {
-      string s = democonf[democonf_sel].second;
-      s.resize(democonf_string_max, ' ');
-      unsigned text_size = 12;
-      unsigned w = 0;
-      unsigned h = 0;
-      COverlayRGB::calculate_boundary_text(s, text_size, w, h);
-
-      int x = 7*SCREEN_W / 8;
-      int y = 7*SCREEN_H / 8;
-
-      COverlayRGB overlay_democonf(SCREEN_W, SCREEN_H);
-      overlay_democonf.alloc_mem_overlay(w, h);
-      overlay_democonf.copy_overlay(bg_overlay,x, y);
-      overlay_democonf.set_text(0, 0, s, text_size, 0x00f4419d);
-      overlay_democonf.print_to_display(x, y);
-    }
+  // Draw background two times for front and back buffer
+  const char *titles[] = {
+    "CNN - GoogLeNet",
+    "Object Class Identification",
+  };
+  for (int i = 0; i < 2; ++i) {
+    bg_overlay.print_to_display(0, 0);
+    print_demo_title(bg_overlay, titles);
+    swap_buffer();
   }
 
-  g_should_stop = true;
-  pthread_join(hwacc_thread, NULL);
-  dmp::util::shutdown();
+  int exit_code = -1;
+  int image_nr = 0;
+  bool pause = false;
+  std::vector<float> network_output;
+  // Enter main loop
+  while (exit_code == -1) {
+    // If not pause, decode next JPEG image and do pre-processing
+    if (!pause) {
+      decode_jpg_file(image_names[image_nr], imgView, IMAGE_W, IMAGE_H);
+      overlay_input.convert_to_overlay_pixel_format(imgView, IMAGE_W * IMAGE_H);
+      preproc_image(imgView, imgProc, IMAGE_W, IMAGE_H, -128.0, -128.0, -128.0,
+                    1.0, true);
+      ++image_nr;
+      image_nr %= num_images;
+    }
+
+    // Run network in HW
+    memcpy(network.get_network_input_addr_cpu(), imgProc, IMAGE_W * IMAGE_H * 6);
+    network.RunNetwork();
+
+    // Handle output from HW
+    network.get_final_output(network_output);
+    overlay_input.print_to_display((SCREEN_W - IMAGE_W) / 2, 185);
+
+    // Print identification result to screen
+    // print_result and catrank functions are defined in util_draw.cpp
+    print_result(catstr_vec, (SCREEN_W / 5), IMAGE_W + 245,
+                 catrank(&network_output.front()), bg_overlay);
+
+    // Output HW processing times
+    int conv_time_tot = network.get_conv_usec();
+    print_conv_time(bg_overlay, 7 * SCREEN_H / 8, conv_time_tot, conv_freq);
+    int fc_time_tot = network.get_fc_usec();
+    print_fc_time(bg_overlay, 7 * SCREEN_H / 8 + 28, fc_time_tot, fc_freq);
+    print_total_time(bg_overlay, 7 * SCREEN_H / 8 + 56, conv_time_tot + fc_time_tot);
+
+    swap_buffer();
+
+    handle_keyboard_input(exit_code, pause);
+  }
+
+  shutdown();
 
   return exit_code;
 }
